refactor sql interactions to repository
